requirements.txt
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv
pyodbc
alembic
openai>=1.35.0
python-dotenv>=1.0.1
django 
djangorestframework==3.16.1             # REST API toolkit
djangorestframework-simplejwt==5.5.1    # JWT auth for DRF
django-environ==0.11.2                  # Environment variables
requests==2.32.3                        
paramiko
beautifulsoup4
environ
django-cors-headers
PyPDF2